vatrade-engine/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Trading Libraries
python-binance==1.0.19
ccxt==4.4.26

# Async & Background Tasks
celery==5.4.0
redis==5.2.0

# HTTP Client
httpx==0.28.0
aiohttp==3.11.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Logging & Monitoring
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
